fix: #15 reflect code review
