+++ src/controllers/missionController.ts
@@ -5,6 +5,7 @@ import { fail, success } from '../constants/response';
 import { slackMessage } from '../modules/slackMessage';
 import { sendMessageToSlack } from '../modules/slackAPI';
 import { missionService } from '../service';
+import dateValidator from '../modules/dateValidator';
 
 /**
  *  @route GET /mission/daily/:date
@@ -15,13 +16,13 @@ const getDailyMission = async (req: Request, res: Response) => {
   try {
     const userId = req.body.userId;
     const actionDate = req.params.date;
+    await dateValidator.validateDate(actionDate);
     const data = await missionService.getDailyMission(userId, actionDate);
-    if (data === 400) {
-      return res.status(statusCode.BAD_REQUEST).send(fail(statusCode.BAD_REQUEST, message.INVALID_DATE_TYPE));
-    }
     return res.status(statusCode.OK).send(success(statusCode.OK, message.READ_DAILY_MISSION_SUCCESS, data));
   } catch (error) {
-    console.log(error);
+    if (error === 400) {
+      res.status(statusCode.BAD_REQUEST).send(fail(statusCode.BAD_REQUEST, message.INVALID_DATE_TYPE));
+    }
     const errorMessage: string = slackMessage(req.method.toUpperCase(), req.originalUrl, error, req.body.user?.id);
     sendMessageToSlack(errorMessage);
     res.status(statusCode.INTERNAL_SERVER_ERROR).send(fail(statusCode.INTERNAL_SERVER_ERROR, message.INTERNAL_SERVER_ERROR));
