+++ src/service/index.ts
@@ -1,2 +1,2 @@
 export { default as authService } from './authService';
-export { default as missionService} from './missionService';
+export { default as missionService } from './missionService';
